@@ -6,13 +6,13 @@ ext.universalBase = 8000
 
 android {
     namespace 'is.xyz.mpv'
-    compileSdkVersion 33
+    compileSdk 34
 
     defaultConfig {
         minSdkVersion 21
         targetSdkVersion 33
-        versionCode 33
-        versionName "2023-11-30-release"
+        versionCode 34
+        versionName "2024-04-09-release"
         vectorDrawables.useSupportLibrary = true
     }
 
@@ -29,6 +29,7 @@ android {
 
     buildFeatures {
         viewBinding = true
+        buildConfig = true
     }
 
     splits {
@@ -72,5 +73,5 @@ dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation 'androidx.appcompat:appcompat:1.6.1'
     implementation 'androidx.recyclerview:recyclerview:1.3.2'
-    implementation 'androidx.media:media:1.6.0'
+    implementation 'androidx.media:media:1.7.0'
 }
