@@ -1,12 +1,12 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
-    ext.kotlin_version = '1.8.22'
+    ext.kotlin_version = '1.9.23'
     repositories {
         mavenCentral()
         google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:8.1.4'
+        classpath 'com.android.tools.build:gradle:8.3.1'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
@@ -18,9 +18,10 @@ allprojects {
     }
 }
 
+// Use `./gradlew wrapper` to re-generate it
 tasks {
     wrapper {
-        gradleVersion = "8.0"
+        gradleVersion = "8.5"
         distributionType = Wrapper.DistributionType.BIN
     }
 }
