@@ -6,6 +6,7 @@ import android.database.Cursor;
 import android.net.Uri;
 import android.os.Build;
 import android.provider.DocumentsContract;
+import android.util.Log;
 
 import androidx.annotation.NonNull;
 import androidx.loader.content.AsyncTaskLoader;
@@ -18,7 +19,7 @@ import java.util.List;
 
 /**
  * An implementation of the picker that operates on a document tree.
- *
+ * <br>
  * See also:
  * - https://developer.android.com/training/data-storage/shared/documents-files
  * - https://developer.android.com/reference/android/provider/DocumentsContract.Document#MIME_TYPE_DIR
@@ -31,10 +32,13 @@ public class DocumentPickerFragment extends AbstractFilePickerFragment<Uri> {
     // grab additional info for free afterwards. This is not the case with the documents API so we
     // have to work around it.
     final HashMap<Uri, Document> mLastRead;
+    // maps document ID of directories to parent path
+    final HashMap<String, Uri> mParents;
 
     public DocumentPickerFragment(@NonNull Uri root) {
         mRoot = root;
         mLastRead = new HashMap<>();
+        mParents = new HashMap<>();
     }
 
     /**
@@ -108,30 +112,24 @@ public class DocumentPickerFragment extends AbstractFilePickerFragment<Uri> {
 
     @NonNull
     @Override
-    public Uri toUri(@NonNull Uri path) {
-        return path;
-    }
-
-    @NonNull
-    @Override
     public Uri getParent(@NonNull Uri from) {
-        Document doc = mLastRead.get(from);
-        if (doc != null) {
-            return doc.parent;
-        }
-        // This is not supposed to happen
+        String docId = DocumentsContract.getDocumentId(from);
+        Uri parent = mParents.get(docId);
+        if (parent != null)
+            return parent;
+        Log.e(TAG, "getParent() has not seen this document before");
         return getRoot();
     }
 
     @NonNull
     @Override
-    public String getFullPath(@NonNull Uri path) {
+    public String pathToString(@NonNull Uri path) {
         return path.toString();
     }
 
     @NonNull
     @Override
-    public Uri getPath(@NonNull String path) {
+    public Uri pathFromString(@NonNull String path) {
         return Uri.parse(path);
     }
 
@@ -148,7 +146,7 @@ public class DocumentPickerFragment extends AbstractFilePickerFragment<Uri> {
         final Uri currentPath = mCurrentPath;
 
         // totally makes sense!
-        final String docId = mCurrentPath.equals(root) ? DocumentsContract.getTreeDocumentId(currentPath) :
+        final String docId = currentPath.equals(root) ? DocumentsContract.getTreeDocumentId(currentPath) :
                 DocumentsContract.getDocumentId(currentPath);
         final Uri childUri = DocumentsContract.buildChildDocumentsUriUsingTree(root, docId);
 
@@ -157,7 +155,7 @@ public class DocumentPickerFragment extends AbstractFilePickerFragment<Uri> {
                 DocumentsContract.Document.COLUMN_MIME_TYPE,
                 DocumentsContract.Document.COLUMN_DISPLAY_NAME,
         };
-        return new AsyncTaskLoader<List<Uri>>(requireContext()) {
+        return new AsyncTaskLoader<>(requireContext()) {
             @Override
             public List<Uri> loadInBackground() {
                 final ContentResolver contentResolver = getContext().getContentResolver();
@@ -170,12 +168,19 @@ public class DocumentPickerFragment extends AbstractFilePickerFragment<Uri> {
                 final int i1 = c.getColumnIndex(cols[0]), i2 = c.getColumnIndex(cols[1]), i3 = c.getColumnIndex(cols[2]);
                 while (c.moveToNext()) {
                     // TODO later: support FileFilter equivalent here
+                    final String docId = c.getString(i1);
+                    final boolean isDir = c.getString(i2).equals(DocumentsContract.Document.MIME_TYPE_DIR);
                     files.add(new Document(
-                            DocumentsContract.buildDocumentUriUsingTree(root, c.getString(i1)),
-                            currentPath,
-                            c.getString(i2).equals(DocumentsContract.Document.MIME_TYPE_DIR),
+                            DocumentsContract.buildDocumentUriUsingTree(root, docId),
+                            isDir,
                             c.getString(i3)
                     ));
+                    // There is no generic way to get a parent directory for another directory and this
+                    // can't be solved via mLastRead either, since by the time someone asks getParent()
+                    // we're already inside the new directory. Not to mention that this would be insufficient
+                    // when going back multiple times.
+                    if (isDir)
+                        mParents.put(docId, currentPath);
                 }
                 c.close();
 
@@ -206,13 +211,11 @@ public class DocumentPickerFragment extends AbstractFilePickerFragment<Uri> {
      */
     private static class Document implements Comparable<Document> {
         private final @NonNull Uri uri;
-        private final @NonNull Uri parent;
         private final boolean isDir;
         private final @NonNull String displayName;
 
-        private Document(@NonNull Uri uri, @NonNull Uri parent, boolean dir, @NonNull String name) {
+        private Document(@NonNull Uri uri, boolean dir, @NonNull String name) {
             this.uri = uri;
-            this.parent = parent;
             isDir = dir;
             displayName = name;
         }
@@ -228,13 +231,11 @@ public class DocumentPickerFragment extends AbstractFilePickerFragment<Uri> {
         // Sort directories before files, alphabetically otherwise
         @Override
         public int compareTo(Document other) {
-            if (isDir && !other.isDir) {
-                return -1;
-            } else if (other.isDir && !isDir) {
-                return 1;
-            } else {
-                return displayName.compareToIgnoreCase(other.displayName);
-            }
+            if (isDir != other.isDir)
+                return other.isDir ? 1 : -1;
+            return displayName.compareToIgnoreCase(other.displayName);
         }
     }
+
+    private static final String TAG = "mpv";
 }
