@@ -6,13 +6,18 @@ import android.text.TextWatcher
 import android.view.LayoutInflater
 import android.view.View
 
-class DecimalPickerDialog(
+internal class DecimalPickerDialog(
     private val rangeMin: Double, private val rangeMax: Double
 ) : PickerDialog {
     private lateinit var binding: DialogDecimalBinding
 
     override fun buildView(layoutInflater: LayoutInflater): View {
         binding = DialogDecimalBinding.inflate(layoutInflater)
+
+        // hide extranous UI parts
+        arrayOf(binding.label1, binding.label2, binding.rowSecondary).forEach {
+            it.visibility = View.GONE
+        }
 
         binding.editText.addTextChangedListener(object : TextWatcher {
             override fun beforeTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {}
@@ -25,14 +30,12 @@ class DecimalPickerDialog(
                     binding.editText.setText(valueBounded.toString())
             }
         })
-        binding.btnMinus.setOnClickListener {
+        val onClick = { delta: Double ->
             val value = this.number ?: 0.0
-            this.number = (value - STEP).coerceIn(rangeMin, rangeMax)
+            this.number = (value - delta).coerceIn(rangeMin, rangeMax)
         }
-        binding.btnPlus.setOnClickListener {
-            val value = this.number ?: 0.0
-            this.number = (value + STEP).coerceIn(rangeMin, rangeMax)
-        }
+        binding.btnMinus.setOnClickListener { onClick(-STEP) }
+        binding.btnPlus.setOnClickListener { onClick(STEP) }
 
         return binding.root
     }
