@@ -3,7 +3,7 @@ package `is`.xyz.mpv
 import android.view.LayoutInflater
 import android.view.View
 
-interface PickerDialog {
+internal interface PickerDialog {
     fun buildView(layoutInflater: LayoutInflater): View
 
     fun isInteger(): Boolean // eh....
