@@ -6,7 +6,7 @@ import android.view.View
 import android.widget.SeekBar
 import androidx.annotation.StringRes
 
-class SliderPickerDialog(
+internal class SliderPickerDialog(
     private val rangeMin: Double, private val rangeMax: Double, private val intScale: Int,
     @StringRes private val formatTextRes: Int
 ) : PickerDialog {
