@@ -6,7 +6,7 @@ import android.view.View
 import android.widget.SeekBar
 import kotlin.math.max
 
-class SpeedPickerDialog : PickerDialog {
+internal class SpeedPickerDialog : PickerDialog {
     companion object {
         // Middle point of bar (in progress units)
         private const val HALF = 100.0
